Updates synergy package to version 1.10.0
I was forced to implement some nastiness in order to get everything
working.

1. Updated the patch to use system libs as opposed to bundled libs.
2. Had to add some include files to fix build with qt 5.11 (likely to be
obsolete in a future version, see: https://github.com/symless/synergy-core/pull/6359)
3. Building the tests was disabled without a nice way to enable it. Had
to `add_subdirectory` in the root CMake file.
4. Had to delete the file
`src/test/integtests/arch/ArchInternetTests.cpp`. This file contained a
reference to a macro `ARCH_INTERNET`. This macro however (or even this
string) exists *nowhere else* in the project, causing the build to fail.
Without the file everything is working fine.
5. Had to fix the test CMake files, as they wanted to link against
`synergy` which is a binary that does not export its symbols...

Cheers,
Luca