linux-hardened: Enable intel iommu by default
Previously there was attempt to enable intel IOMMU by default on Arch kernels which was reverted due to issues with intel graphics stack. In linux 4.18 the long standing IOMMU bug that affected intel graphics was fixed[0] therefore we can try enabling it now. At least on linux-hardened. IOMMU is important for security of virtual machines.

[0] https://bugs.freedesktop.org/show_bug.cgi?id=89360