upgpkg: patch 2.7.6-7 (fix patches for real)
remove unneeded local patches and use only upstream patches
where possible.
19599883ffb6a450d2884f081f8ecf68edbed7ee requres a tiny context
fix in a single hunk, so we carry that patch around. Its easy
to verify against the upstream commit.

All changes in all patches are exactly like the upstream changes
not fiddling with mixups and not undoing CVE-2018-6952.

The autoreconf was added so we get a new Makefile that respects
the newly added tests/ed-style so we can ensure it still works.